authentication and authorization with JWT

+++ controllers/api/v1/Post_api.js
@@ -27,8 +27,8 @@ module.exports.index =async function(req,res){
 module.exports.destroy = async function(req,res){
         try{
             let post= await Post.findById(req.params.id);
-
-            //if(post.user == req.user.id){
+            //if the user requesting to delete the post exists
+            if(post.user == req.user.id){
                 post.remove();
 
                 await  Comment.deleteMany({post: req.params.id});
@@ -36,7 +36,13 @@ module.exports.destroy = async function(req,res){
                 return res.json(200,{
                     message :"post and associated comments deleted successfully!"
                 });
-            // }
+            }
+            // if user doesnot exists
+            else{
+                return res.json(401,{
+                    message :"you cannot delete this post"
+                })
+            }
         }catch(err){
             console.log('******',err);
             return res.json(500,{
